新增链表的应用2

+++ hoho-mod-test/src/main/java/com/linx/test/algorithm/practise/SqListPractise.java
@@ -8,14 +8,14 @@ public class SqListPractise {
      * 设计一个算法，以第一个元素为分界线，将所有小于它的元素移到该元素的前面，将所有大于它的元素移动到该元素的后面
      */
     private void move(SqList<Integer> list) {
-        Integer pivot = list.data[0];
+        Integer pivot = (Integer) list.data[0];
         int i = 0, j = list.length -1;
         while (i != j) {
-            while (j > i && list.data[j] > pivot) { //从右边找到第一个比pivot小的数。此处大于时即进入循环，小于则跳出循环
+            while (j > i && ((Integer) list.data[j]) > pivot) { //从右边找到第一个比pivot小的数。此处大于时即进入循环，小于则跳出循环
                 j--;
             }
             list.data[i] = list.data[j];
-            while (i < j && list.data[i] < pivot) {//从左边找到第一个比pivot大的数
+            while (i < j && ((Integer) list.data[i]) < pivot) {//从左边找到第一个比pivot大的数
                 i++;
             }
             list.data[j] = list.data[i];// 两个交换
